A few changes

+++ backend/src/utils/equipo.rs
@@ -1,6 +1,6 @@
 use std::{env, path::PathBuf};
 
-pub fn path_team_img(id: u32) -> Result<PathBuf, Box<dyn std::error::Error>> {
+pub fn path_team_img(id: i64) -> Result<PathBuf, Box<dyn std::error::Error>> {
     let mut route = PathBuf::from(env::var("ROUTE_TO_IMG")?);
     route.push("img_equipo");
     route.push(format!("{id}"));
